requirements.txt
alembic==1.17.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg-binary==3.2.12
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
